requirements.txt
transformers
Pillow
requests
ipython
open3d
torch
matplotlib
numpy
